requirements.txt
pymupdf
sentence-transformers
faiss-cpu
streamlit
openai
numpy
